+++ tests/testthat/test-get_election_data.R
@@ -28,7 +28,7 @@ test_that("get_elections_data works", {
     sample(x = as_date(c("1982-10-28", "1986-06-22", "1989-10-29",
                          "1993-06-06", "1996-03-03", "2000-03-12",
                          "2004-03-14", "2008-03-09", "2011-11-20",
-                         "2015-12-20", "2016-06-26", "2023-07-24")), size = 2,
+                         "2015-12-20", "2016-06-26", "2023-07-24")), size = 1,
            replace = FALSE)
   expect_equal(get_election_data(type_elec = "congress", date = random_dates,
                                  verbose = FALSE) |>
@@ -42,7 +42,7 @@ test_that("get_elections_data works", {
     sample(x = as_date(c("1982-10-28", "1986-06-22", "1989-10-29",
                          "1993-06-06", "1996-03-03", "2000-03-12",
                          "2004-03-14", "2008-03-09", "2011-11-20",
-                         "2015-12-20", "2016-06-26", "2023-07-24")), size = 2,
+                         "2015-12-20", "2016-06-26", "2023-07-24")), size = 1,
            replace = FALSE)
   expect_equal(get_election_data(type_elec = "congress", date = random_dates,
                                  verbose = FALSE) |>